Create and Test Bestowed Magic
Bestowed Magic generation is now working.

Added a MINIMUM bonus type so that the score after all calculations must
be at least this value.

A Hero now has an age field, though there is not yet a generator for it.
The Magic Bestowed has a Old and Kid limitations that set this.

A Hero can have a skill added that takes the guess work about how to
just add a skill without giving it a value.

Super Abilities has a rollAbilities function for easier adding super
abilities.

Skills has a physical skills function for getting all possible physical
skills as well as a function to exclude skills from a list of skills.

Fixed dice rolling for a Test Roll that doesn't specify a result that it
gets a result instead of returning an undefined value.